+++ src/main/java/Controller/CustomerDetailController.java
@@ -1,0 +1,41 @@
+package Controller;
+
+import java.io.*;
+import java.io.OutputStream;
+
+import javax.servlet.RequestDispatcher;
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
+@WebServlet("/customerDetail")
+public class CustomerDetailController extends HttpServlet {
+	private static final long serialVersionUID = 1L;
+	
+	public CustomerDetailController() {
+		super();
+	}
+
+	@Override
+	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
+		getCustomerListProcess(request, response);
+	}
+	
+	@Override
+	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
+	}
+	
+	protected void getCustomerListProcess(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		try {
+			String id = request.getParameter("id");
+			request.setAttribute("id", id);
+			RequestDispatcher dispatcher = request.getRequestDispatcher(request.getContextPath() + "/components/popup/NewFile.jsp");
+	        dispatcher.forward(request, response);
+	        } catch (Exception e) {
+			e.printStackTrace();
+		}
+	}
+}
