+++ src/main/webapp/components/popup/NewFile.jsp
@@ -1,0 +1,109 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+    pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>Insert title here</title>
+</head>
+<body>
+<div class="popupBox display">
+				<svg class="popupExitButton" width="14" height="15" viewBox="0 0 14 15" fill="none" xmlns="http://www.w3.org/2000/svg">
+					<path d="M14 1.91L12.59 0.5L7 6.09L1.41 0.5L0 1.91L5.59 7.5L0 13.09L1.41 14.5L7 8.91L12.59 14.5L14 13.09L8.41 7.5L14 1.91Z" fill="#323232"/>
+				</svg>
+				<div class="popupTitleBox"><h1>입출금 내역</h1></div>
+				<table class="searchTable" id="historyTable">
+					<tr>
+						<th>계좌일자</th>
+						<th>거래유형</th>
+						<th>계좌번호</th>
+						<th>고객 이름</th>
+						<th>입금자명</th>
+						<th>거래처</th>
+						<th>거래 금액</th>
+						<th>계좌 잔액</th>
+					</tr>
+					<tr>
+						<td>2023-06-14</td>
+						<td>급여</td>
+						<td>3333-08-9795700</td>
+						<td>김민재</td>
+						<td>이상준</td>
+						<td>3조</td>
+						<td>9,400,000원</td>
+						<td>26,613,000원</td>
+					</tr>
+					<tr>
+						<td>2023-06-14</td>
+						<td>급여</td>
+						<td>3333-08-9795700</td>
+						<td>김민재</td>
+						<td>이상준</td>
+						<td>3조</td>
+						<td>9,400,000원</td>
+						<td>26,613,000원</td>
+					</tr>
+					<tr>
+						<td>2023-06-14</td>
+						<td>급여</td>
+						<td>3333-08-9795700</td>
+						<td>김민재</td>
+						<td>이상준</td>
+						<td>3조</td>
+						<td>9,400,000원</td>
+						<td>26,613,000원</td>
+					</tr>
+					<tr>
+						<td>2023-06-14</td>
+						<td>급여</td>
+						<td>3333-08-9795700</td>
+						<td>김민재</td>
+						<td>이상준</td>
+						<td>3조</td>
+						<td>9,400,000원</td>
+						<td>26,613,000원</td>
+					</tr>
+					<tr>
+						<td>2023-06-14</td>
+						<td>급여</td>
+						<td>3333-08-9795700</td>
+						<td>김민재</td>
+						<td>이상준</td>
+						<td>3조</td>
+						<td>9,400,000원</td>
+						<td>26,613,000원</td>
+					</tr>
+					<tr>
+						<td>2023-06-14</td>
+						<td>급여</td>
+						<td>3333-08-9795700</td>
+						<td>김민재</td>
+						<td>이상준</td>
+						<td>3조</td>
+						<td>9,400,000원</td>
+						<td>26,613,000원</td>
+					</tr>
+					<tr>
+						<td>2023-06-14</td>
+						<td>급여</td>
+						<td>3333-08-9795700</td>
+						<td>김민재</td>
+						<td>이상준</td>
+						<td>3조</td>
+						<td>9,400,000원</td>
+						<td>26,613,000원</td>
+					</tr>
+					<tr>
+						<td>나는</td>
+						<td>팝업도</td>
+						<td>한다고</td>
+						<td>한 적이</td>
+						<td>없는데</td>
+						<td>왜 나한테</td>
+						<td>시키는</td>
+						<td>거야</td>
+					</tr>
+				</table>
+			</div>
+</body>
+</html>
