+++ src/main/webapp/view/deposit/depositProductList/depositProductList.js
@@ -1,4 +1,5 @@
 selectOneItem('depositBalance');
+selectMultiItemsWithDirectInput("depositStartDate");
 selectOneItem('depositType');
 
 const searchTableRows = document.querySelectorAll('#depositSearchTable tr :not(th)');
@@ -14,3 +15,8 @@ searchTableRows.forEach((item) => {
 popupExitButton.addEventListener('click', () => {
 	popupBox.classList.toggle('display');
 });
+
+setYearSelect();
+setMonthSelect();
+setDaySelect(true);
+changeDate();
