+++ src/main/webapp/view/employee/empList/empList.css
@@ -1,5 +1,6 @@
 @charset "UTF-8";
 
+
 .innerSelectBox {
 	margin-right: 200px;
 }
