+++ src/main/webapp/view/login/login.css
@@ -11,7 +11,6 @@
 }
 .loginOuterbox{
 	width: 508px;
-	height: 365px;
 	border-radius: 12px;
 	gap: 20px;
 	background: white;
@@ -20,19 +19,21 @@
 .loginInnerbox{
 	width: 100%;
 	height: fit-content;
-	padding: 40px;
+	padding: 20px 40px;
 	box-sizing: border-box;
 	margin: auto;
 }
 .loginTitleBox{
 	width: 100%;
-	height: 39px;
 	text-align: center;
 	background: balck;
 	font-size: 24px;
+	font-weight: bold;
 	display: flex;
 	flex-direction: column;
 	justify-content: flex-start;
+	padding-top: 20px;
+	padding-bottom: 40px;
 }
 .loginBodyBox{
 	width: 100%;
@@ -70,11 +71,11 @@
 .loginSignupButton{
 	margin-top: auto;
 	width: 100%;
-	height: 19px;
 	text-align: center;
 	font-size: 14px;
 	text-decoration: none;
 	color: #8B95A1;
+	padding: 10px 0;
 }
 .makeButtonGreen{
 	background: #2DB400;
